rename some files to match exports
